GNCF class but with feature input TODO: The NGCF paper first embeds users and items in the same space and then applies the GNN.
This would be more flexible allowing for different dimensions in items and features but what would it mean for the Data.x graph created by the provider?

We could still use the same graph but manually apply an extra embedding step before the GNN convolutions so long as we know where the items stop and the users begin in the `x` tensor. With different dimensions we can't even have an x vector...